clean up front end
