+++ ngapp/src/app/admin/add-content/add-content.component.html
@@ -1,111 +1,20 @@
 <div class="mainContainer">
-  <div class="headerTitle">Add Prompt Data to DB</div>
-  <!-- Select Content Type-->
-  <mat-form-field appearance="fill">
-    <mat-label>Select Prompt Type</mat-label>
-    <mat-select [(ngModel)]="selectedPromptGenerator" >
-      <mat-option value="general">General Prompts</mat-option>
-      <mat-option value="proverb">Proverb Prompts</mat-option>
-      <mat-option value="exam">Exam Prompts</mat-option>
-      <mat-option value="lara">LARA Prompts</mat-option>
-      <mat-option value="combination">Combination Prompts</mat-option>
-      <mat-option value="partOfSpeech">Part-of-Speech Prompts</mat-option>
-    </mat-select>
-  </mat-form-field>
-
-  <app-prompt-data-table [hidden]="!selectedPromptGenerator" #promptDataTable [selectedPromptGenerator]="selectedPromptGenerator" ></app-prompt-data-table>
-
-  <!-- Add part of speech data -->
-  <div *ngIf="selectedPromptGenerator == 'partOfSpeech'">
-    <div class="posFields">
-      <mat-form-field appearance="fill">
-        <!-- Part of speech -->
-        <mat-label>Part of Speech</mat-label>
-        <mat-select [(ngModel)]="selectedPOS">
-          <mat-option value="noun">{{ ts.l.noun }}</mat-option>
-          <mat-option value="verb">{{ ts.l.verb }}</mat-option>
-          <mat-option value="adjective">{{ ts.l.adjective }}</mat-option>
-          <mat-option value="preposition">{{ ts.l.preposition }}</mat-option>
-          <mat-option value="adverb">{{ ts.l.adverb }}</mat-option>
-          <mat-option value="pronoun">{{ ts.l.pronoun }}</mat-option>
-          <mat-option value="article">{{ ts.l.article }}</mat-option>
-          <mat-option value="conjunction">{{ ts.l.conjunction }}</mat-option>
-        </mat-select>
-      </mat-form-field>
-      <div style="width: 50%; padding: 10px"> 
-        POS entries can be added one at a time, or you can add several at once with each entry on a new line (enter key after each one).
-        Enter the word and the translation separated by a ':' Example:
-      </div>
-      <div> an t-aerfort : the airport</div>
-      <div> an teach : the house</div>
-      <mat-form-field appearance="fill" class="inputText">
-        <mat-label>Word/Translation Entry</mat-label>
-        <textarea matInput [(ngModel)]="posInput"></textarea>
-      </mat-form-field>
-      <div *ngIf="errorMessages.length > 0" class="errorText"> <div *ngFor="let message of errorMessages">{{ message }} </div></div>
-      <button mat-raised-button (click)="addPOSContent()" class="viewDataBtn"> Add </button>
-    </div>
-    <!-- Show part of speech data -->
-    <!-- <button mat-raised-button (click)="toggleTable('pos')" class="viewDataBtn">
-      {{ this.posDataTable.hideTable ? "Show Data" : "Hide Data" }}
-    </button> -->
+  <div class="propmtSelectionContainer">
+    <div class="headerTitle">Select Prompt Generator:</div>
+    <br />
+    <!-- Select Prompt Generator -->
+    <mat-form-field appearance="fill">
+      <mat-label>Prompt Generator</mat-label>
+      <mat-select [(ngModel)]="selectedPromptGenerator">
+        <mat-option value="general">General Prompts</mat-option>
+        <mat-option value="proverb">Proverb Prompts</mat-option>
+        <mat-option value="exam">Exam Prompts</mat-option>
+        <mat-option value="lara">LARA Prompts</mat-option>
+        <mat-option value="combination">Combination Prompts</mat-option>
+        <mat-option value="partOfSpeech">Part-of-Speech Prompts</mat-option>
+      </mat-select>
+    </mat-form-field>
   </div>
 
-  <!-- Add prompt data -->
-  <div *ngIf="selectedPromptGenerator !== 'partOfSpeech'">
-    <div class="posFields">
-      <!-- Choose dialect -->
-      <mat-form-field appearance="fill" *ngIf="selectedPromptGenerator == 'proverb'" >
-        <mat-label>{{ ts.l.choose_dialect }}</mat-label>
-        <mat-select [(ngModel)]="selectedDialect">
-          <mat-option value="munster">{{ this.ts.l.munster }}</mat-option>
-          <mat-option value="connacht">{{ this.ts.l.connacht }}</mat-option>
-          <mat-option value="ulster">{{ this.ts.l.ulster }}</mat-option>
-        </mat-select>
-      </mat-form-field>
-      <!-- Select level -->
-      <mat-form-field appearance="fill" *ngIf=" selectedPromptGenerator == 'combination' || selectedPromptGenerator == 'exam' " >
-        <mat-label>{{ ts.l.choose_level }}</mat-label>
-        <mat-select [(ngModel)]="selectedLevel">
-          <mat-option value="jc"> {{ this.ts.l.level_preferences[0] }} </mat-option>
-          <mat-option value="lcol"> {{ this.ts.l.level_preferences[1] }} </mat-option>
-          <mat-option value="lchl"> {{ this.ts.l.level_preferences[2] }} </mat-option>
-        </mat-select>
-      </mat-form-field>
-      <!-- Input prompt text -->
-      <div *ngIf="selectedPromptGenerator !== 'combination'" style="width: 50%; padding: 10px"> 
-        Prompt texts can be added one at a time, or you can add several at once with each prompt on a new line (enter key after each one)
-      </div>
-      <mat-form-field appearance="fill" *ngIf="selectedPromptGenerator !== 'combination'" class="inputText" >
-        <mat-label>Prompt Text</mat-label>
-        <textarea matInput [(ngModel)]="promptText"></textarea>
-      </mat-form-field>
-      <div *ngIf="selectedPromptGenerator == 'combination'" style="width: 50%; padding: 10px"> 
-        Please only add one character, setting, and theme at a time
-      </div>
-      <!-- Input prompt character -->
-      <mat-form-field appearance="fill" *ngIf="selectedPromptGenerator == 'combination'" class="inputText"  >
-        <mat-label>{{ ts.l.character }}</mat-label>
-        <textarea matInput [(ngModel)]="selectedCharacter"></textarea>
-      </mat-form-field>
-      <!-- Input prompt setting -->
-      <mat-form-field appearance="fill" *ngIf="selectedPromptGenerator == 'combination'" class="inputText"  >
-        <mat-label>{{ ts.l.setting }}</mat-label>
-        <textarea matInput [(ngModel)]="selectedSetting"></textarea>
-      </mat-form-field>
-      <!-- Input prompt theme -->
-      <mat-form-field appearance="fill" *ngIf="selectedPromptGenerator == 'combination'" class="inputText"  >
-        <mat-label>{{ ts.l.theme }}</mat-label>
-        <textarea matInput [(ngModel)]="selectedTheme"></textarea>
-      </mat-form-field>
-
-      <div *ngIf="errorMessages.length > 0" class="errorText"> <div *ngFor="let message of errorMessages">{{ message }} </div></div>
-      <button mat-raised-button (click)="addPromptContent()" class="viewDataBtn" > Add </button>
-    </div>
-  </div>
+  <app-prompt-data-table [hidden]="!selectedPromptGenerator" [selectedPromptGenerator]="selectedPromptGenerator"></app-prompt-data-table>
 </div>
-
-<!-- Data Tables -->
-<!-- <app-prompt-data-table #promptDataTable [data]="promptDataSource" ></app-prompt-data-table> -->
-
-<!-- <app-pos-data-table #posDataTable [data]="posDataSource"></app-pos-data-table> -->
