+++ ngapp/src/app/admin/add-content/prompt-data-table/prompt-data-table.component.scss
@@ -1,99 +1,72 @@
-// table {
-//     width: 100%;
-//     text-align: left;
-// }
 
-// td, th {
-//     text-align: left !important;
-//     vertical-align: middle !important;
-// }
+h3 {
+  text-align: left;
+  padding-top: 20px;
+  text-decoration: underline;
+}
 
-// .mat-sort-header-container {
-//     justify-content: space-evenly !important;
-// }
+.tableButtons {
+  width: 100%;
+  margin: auto;
+  text-align: right;
+  margin-bottom: 10px;
+}
 
-// .example-container {
-//     overflow-x: scroll;
-//     max-width: 60px;
-//     text-align: left !important;
-//   }
+.button-add-row {
+  background: var(--scealai-green);
+  margin-right: 5px;
+}
 
-// .mat-mdc-form-field {
-//     font-size: 14px;
-//     width: 100%;
-// }
+.button-remove-rows {
+  background: red;
+}
 
-p {
-    font-family: Lato;
-  }
-  table {
-    width: 90%;
-    margin: auto;
-    margin-top: 25px;
-  }
-  .btn-edit {
-    text-align: right;
-  }
-  
-  .table-header {
-    width: 90%;
-    margin: auto;
-    text-align: right;
-    margin-bottom: 10px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-  
-  .button-add-row {
-    background: var(--scealai-green);
-    color: #ffffff;
-    margin-right: 5px;
-  }
-  
-  .button-remove-rows {
-    background: red;
-    color: #ffffff;
-    margin-right: 5px;
-  }
-  
-  .button-remove {
-    color: red;
-  }
-  
-  ::ng-deep .mat-form-field-infix {
-    width: 100% !important;
-  }
-  
-  .form-input {
-    padding-right: 5px;
-  }
+.multipleRowInputContainer {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
 
-  .multipleRowInputContainer {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    align-items: center;
-  }
+.multipleRowInputInstructions{
+  width: 40%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-evenly;
+}
 
-  .multipleRowInputInstructions{
-    width: 40%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
-  }
-  .multipleRowInput {
-    width: 90%;
-    text-align: left;
-  }
+.multipleRowInput {
+  width: 90%;
+  text-align: left;
+}
 
+ul {
+  text-align: left;
+}
 
+table {
+  width: 90%;
+  margin: auto;
+  margin-top: 25px;
+}
+
+.btn-edit {
+  text-align: right;
+}
+
+::ng-deep .mat-form-field-infix {
+  width: 100% !important;
+}
+
+.form-input {
+  padding-right: 5px;
+}
 
 .mat-mdc-checkbox.mat-accent {
-    --mdc-checkbox-selected-checkmark-color: #fff;
-    --mdc-checkbox-selected-focus-icon-color: var(--scealai-green);
-    --mdc-checkbox-selected-hover-icon-color: var(--scealai-green);
-    --mdc-checkbox-selected-icon-color: var(--scealai-green);
-    --mdc-checkbox-selected-pressed-icon-color: var(--scealai-green);
-  }
+  --mdc-checkbox-selected-checkmark-color: #fff;
+  --mdc-checkbox-selected-focus-icon-color: var(--scealai-green);
+  --mdc-checkbox-selected-hover-icon-color: var(--scealai-green);
+  --mdc-checkbox-selected-icon-color: var(--scealai-green);
+  --mdc-checkbox-selected-pressed-icon-color: var(--scealai-green);
+}
