+++ ngapp/src/app/admin/admin-dashboard/admin-dashboard.component.html
@@ -26,7 +26,7 @@
         Feature Stats
       </div>
       <div class="adminBtnContainer" routerLink="add-content">
-        Add Content
+        Cén Scéal Data
       </div>
       <a class="adminBtnContainer" href="/api/status">
         Backend Monitor
